[debops.ansible_plugins] Add 'ldap_attrs' module
The 'ldap_attrs' Ansible module is a replacement for the 'ldap_attr'
module created by Alexander Korinek. It supports management of multiple
LDAP attributes at once, which solves several issues with LDAP
configuration.

At the moment the 'ldap_attrs' module is not part of Ansible core,
therefore it will be included in the 'debops.ansible_plugins' role for
the forseeable future.

Ref: https://github.com/ansible/ansible/pull/31664