[debops.ansible_plugins] Support for X-ORDERED ext
This patch adds support for automatically prepending the X-ORDERED index
numbers to the values in the list format, in all object attributes. This
is useful for maintaining the LDAP Access Control List (olcAccess)
without the need to renumber all of its entries every time an entry is
added or removed.