[debops.ansible_plugins] Expand simple strings
This patch implements expansion of simple string list items into YAML
dictionaries in the 'parse_kv_items' filter plugin. This can be used to,
for example, convert a list of hosts in Ansible inventory group to
a merged list of entries which can then be augmented with options and
other parameters.